Adding my personal blacklist
This blacklist is the one I use for the family Internet and is the basis of the research into the ever-growing pervasive advertising and tracking industry. Where possible, all entries have been mapped to understand where they fit into the overall ecosystem and if they are serving tracking-related content or tools.